AbstractFloat; Merge ASC into reconstruct_trajectory
