+++ src/TaD.jl
@@ -16,6 +16,5 @@ module TaD
 include("trajectory.jl")
 include("bspline.jl")
 include("models.jl")
-include("asc.jl")
 
 end # module
